Пакет obj переименован в gameobject
