+++ src/main/java/legui/MapBig.java
@@ -1,8 +1,5 @@
 package legui;
 
-import com.google.gson.internal.$Gson$Preconditions;
-import tow.engine.obj.ObjFactory;
-
 import java.io.*;
 import java.util.LinkedList;
 import java.util.List;
