+++ src/main/java/tow/engine/gameobject/GameObject.java
@@ -1,4 +1,4 @@
-package tow.engine.obj;
+package tow.engine.gameobject;
 
 import java.util.*;
