+++ src/main/java/tow/engine/gameobject/components/Movement.java
@@ -1,7 +1,7 @@
-package tow.engine.obj.components;
+package tow.engine.gameobject.components;
 
 import tow.engine.Global;
-import tow.engine.obj.QueueComponent;
+import tow.engine.gameobject.QueueComponent;
 
 import java.util.Arrays;
 import java.util.List;
