+++ src/main/java/tow/engine/gameobject/components/Position.java
@@ -1,9 +1,9 @@
-package tow.engine.obj.components;
+package tow.engine.gameobject.components;
 
 import tow.engine.Global;
 import tow.engine.Vector2;
 import tow.engine.map.Location;
-import tow.engine.obj.QueueComponent;
+import tow.engine.gameobject.QueueComponent;
 
 import java.util.Arrays;
 import java.util.List;
