+++ src/main/java/tow/engine/gameobject/components/particles/ParticlesGeometry.java
@@ -1,11 +1,10 @@
-package tow.engine.obj.components.particles;
+package tow.engine.gameobject.components.particles;
 
 import tow.engine.Global;
 import tow.engine.Vector2;
-import tow.engine.resources.textures.Texture;
 import org.lwjgl.opengl.GL11;
 
-public class ParticlesTexture extends Particles {
+public class ParticlesGeometry extends Particles {
 
     @Override
     public void destroy() { }
@@ -16,7 +15,6 @@ public class ParticlesTexture extends Particles {
 
         for (Part part : parts){
             part.color.bind();
-            part.textureHandler.texture.bind();
             Vector2<Integer> relativePosition = Global.location.camera.toRelativePosition(new Vector2((int) part.x, (int) part.y));
 
             double defaultX, defaultY;
@@ -31,18 +29,16 @@ public class ParticlesTexture extends Particles {
                 defaultY = relativePosition.y;
             }
 
-            GL11.glBegin(GL11.GL_QUADS);
-            GL11.glTexCoord2f(0,0);
+            int glBeginType = GL11.GL_QUADS;
+            if (part.type.equals(Part.Type.FILL)) glBeginType = GL11.GL_QUADS;
+            if (part.type.equals(Part.Type.HOLLOW)) glBeginType = GL11.GL_LINE_LOOP;
+
+            GL11.glBegin(glBeginType);
             GL11.glVertex2f((float) (defaultX-part.width/2), (float) (defaultY-part.height/2));
-            GL11.glTexCoord2f(1,0);
             GL11.glVertex2f((float) (defaultX+part.width/2), (float) (defaultY-part.height/2));
-            GL11.glTexCoord2f(1,1);
             GL11.glVertex2f((float) (defaultX+part.width/2), (float) (defaultY+part.height/2));
-            GL11.glTexCoord2f(0,1);
             GL11.glVertex2f((float) (defaultX-part.width/2), (float) (defaultY+part.height/2));
             GL11.glEnd();
-            Texture.unbind();
         }
     }
-
 }
