+++ src/main/java/tow/engine/input/mouse/MouseCursor.java
@@ -4,10 +4,10 @@ import org.lwjgl.BufferUtils;
 import tow.engine.Global;
 import tow.engine.Vector2;
 import tow.engine.image.TextureHandler;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.components.Position;
-import tow.engine.obj.components.render.Rendering;
-import tow.engine.obj.components.render.Sprite;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.components.Position;
+import tow.engine.gameobject.components.render.Rendering;
+import tow.engine.gameobject.components.render.Sprite;
 
 import java.nio.DoubleBuffer;
