+++ src/main/java/tow/engine/map/Camera.java
@@ -2,8 +2,8 @@ package tow.engine.map;
 
 import tow.engine.Global;
 import tow.engine.Vector2;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.components.Position;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.components.Position;
 
 public class Camera {
