+++ src/main/java/tow/engine/map/DepthVector.java
@@ -1,9 +1,9 @@
 package tow.engine.map;
 
 import tow.engine.Vector2;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
 
 import java.util.ArrayList;
