+++ src/main/java/tow/game/client/NetGameRead.java
@@ -9,7 +9,7 @@ import tow.engine.map.Background;
 import tow.engine.map.Border;
 import tow.engine.map.Location;
 import tow.engine.net.client.Message;
-import tow.engine.obj.GameObject;
+import tow.engine.gameobject.GameObject;
 import tow.game.client.map.Box;
 import tow.game.client.map.MapObject;
 import tow.game.client.map.Wall;
