+++ src/main/java/tow/game/client/map/Wall.java
@@ -3,7 +3,7 @@ package tow.game.client.map;
 import tow.engine.Global;
 import tow.engine.logger.Logger;
 import tow.engine.image.TextureHandler;
-import tow.engine.obj.components.Collision;
+import tow.engine.gameobject.components.Collision;
 
 public class Wall extends MapObject {
