+++ src/main/java/tow/game/client/particles/Pixel.java
@@ -1,7 +1,7 @@
 package tow.game.client.particles;
 
-import tow.engine.obj.components.particles.Part;
-import tow.engine.obj.components.particles.ParticlesGeometry;
+import tow.engine.gameobject.components.particles.Part;
+import tow.engine.gameobject.components.particles.ParticlesGeometry;
 import tow.engine.image.Color;
 
 public class Pixel extends ParticlesGeometry {
