+++ src/main/java/tow/game/client/tanks/Tank.java
@@ -2,15 +2,15 @@ package tow.game.client.tanks;
 
 import org.liquidengine.legui.component.Label;
 import tow.engine.Global;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.ObjFactory;
-import tow.engine.obj.components.Follower;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
-import tow.engine.obj.components.particles.Particles;
-import tow.engine.obj.components.render.Animation;
-import tow.engine.obj.components.render.GUIElement;
-import tow.engine.obj.components.render.Rendering;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.ObjFactory;
+import tow.engine.gameobject.components.Follower;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
+import tow.engine.gameobject.components.particles.Particles;
+import tow.engine.gameobject.components.render.Animation;
+import tow.engine.gameobject.components.render.GUIElement;
+import tow.engine.gameobject.components.render.Rendering;
 import tow.game.client.ClientData;
 import tow.game.client.GameSetting;
 import tow.game.client.particles.Explosion;
