+++ src/main/java/tow/game/client/tanks/enemy/Enemy.java
@@ -3,13 +3,13 @@ package tow.game.client.tanks.enemy;
 import tow.engine.Global;
 import tow.engine.image.TextureHandler;
 import tow.engine.image.TextureManager;
-import tow.engine.obj.ObjFactory;
-import tow.engine.obj.components.Follower;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
-import tow.engine.obj.components.render.Animation;
-import tow.engine.obj.components.render.Rendering;
-import tow.engine.obj.components.render.Sprite;
+import tow.engine.gameobject.ObjFactory;
+import tow.engine.gameobject.components.Follower;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
+import tow.engine.gameobject.components.render.Animation;
+import tow.engine.gameobject.components.render.Rendering;
+import tow.engine.gameobject.components.render.Sprite;
 import tow.game.client.ClientData;
 import tow.game.client.tanks.Tank;
