+++ src/main/java/tow/game/client/tanks/equipment/bullet/BMass.java
@@ -3,8 +3,8 @@ package tow.game.client.tanks.equipment.bullet;
 import tow.engine.Global;
 import tow.engine.logger.Logger;
 import tow.engine.Loader;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.components.Movement;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.components.Movement;
 import tow.engine.setting.ConfigReader;
 import tow.game.client.map.Wall;
 import tow.game.client.tanks.enemy.EnemyArmor;
