+++ src/main/java/tow/game/client/tanks/player/Bullet.java
@@ -5,14 +5,14 @@ import tow.engine.Global;
 import tow.engine.image.TextureHandler;
 import tow.engine.image.TextureManager;
 import tow.engine.map.Border;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.ObjFactory;
-import tow.engine.obj.components.Collision;
-import tow.engine.obj.components.CollisionDirect;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
-import tow.engine.obj.components.particles.Particles;
-import tow.engine.obj.components.render.Sprite;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.ObjFactory;
+import tow.engine.gameobject.components.Collision;
+import tow.engine.gameobject.components.CollisionDirect;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
+import tow.engine.gameobject.components.particles.Particles;
+import tow.engine.gameobject.components.render.Sprite;
 import tow.engine.setting.ConfigReader;
 import tow.game.client.ClientData;
 import tow.game.client.GameSetting;
