+++ src/main/java/tow/game/client/tanks/player/Player.java
@@ -2,14 +2,14 @@ package tow.game.client.tanks.player;
 
 import org.liquidengine.legui.component.Label;
 import tow.engine.Global;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.ObjFactory;
-import tow.engine.obj.components.Follower;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
-import tow.engine.obj.components.render.Animation;
-import tow.engine.obj.components.render.GUIElement;
-import tow.engine.obj.components.render.Rendering;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.ObjFactory;
+import tow.engine.gameobject.components.Follower;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
+import tow.engine.gameobject.components.render.Animation;
+import tow.engine.gameobject.components.render.GUIElement;
+import tow.engine.gameobject.components.render.Rendering;
 import tow.game.client.ClientData;
 import tow.game.client.GameSetting;
 import tow.game.client.tanks.Effect;
