+++ src/main/java/tow/game/client/tanks/player/PlayerController.java
@@ -4,10 +4,10 @@ import tow.engine.Global;
 import tow.engine.Loader;
 import tow.engine.Vector2;
 import tow.engine.map.Border;
-import tow.engine.obj.GameObject;
-import tow.engine.obj.components.Collision;
-import tow.engine.obj.components.Movement;
-import tow.engine.obj.components.Position;
+import tow.engine.gameobject.GameObject;
+import tow.engine.gameobject.components.Collision;
+import tow.engine.gameobject.components.Movement;
+import tow.engine.gameobject.components.Position;
 import tow.game.client.ClientData;
 import tow.game.client.map.Box;
 import tow.game.client.map.Wall;
